:recycle: Rename to walk
